+++ tools/level_editor/preferences_dialog.vala
@@ -13,7 +13,7 @@ public class PreferencesDialog : Gtk.Dialog
 	// Data
 	public LevelEditorApplication _application;
 
-	// Widgets
+	// Document page.
 	public ColorButtonVector3 _grid_color_button;
 	public ColorButtonVector3 _grid_disabled_color_button;
 	public ColorButtonVector3 _axis_x_color_button;
@@ -21,18 +21,22 @@ public class PreferencesDialog : Gtk.Dialog
 	public ColorButtonVector3 _axis_z_color_button;
 	public ColorButtonVector3 _axis_selected_color_button;
 	public EntryDouble _gizmo_size_spin_button;
-	public Grid _document_grid;
+	public PropertyGridSet _document_set;
 
+	// Viewport page.
 	public EntryDouble _level_autosave_spin_button;
-	public Grid _viewport_grid;
+	public PropertyGridSet _viewport_set;
 
+	// System page.
 	public EntryDouble _log_delete_after_days;
 	public EntryDouble _console_max_lines;
-	public Grid _system_grid;
-
-	public PropertyGridSet _document_set;
-	public PropertyGridSet _viewport_set;
 	public PropertyGridSet _system_set;
+
+	// External Tools page.
+	public AppChooserButton _lua_external_tool_button;
+	public AppChooserButton _image_external_tool_button;
+	public PropertyGridSet _external_tools_set;
+
 	public Gtk.Notebook _notebook;
 
 	public PreferencesDialog(LevelEditorApplication app)
@@ -51,7 +55,10 @@ public class PreferencesDialog : Gtk.Dialog
 		_viewport_set.border_width = 12;
 		_system_set = new PropertyGridSet();
 		_system_set.border_width = 12;
+		_external_tools_set = new PropertyGridSet();
+		_external_tools_set.border_width = 12;
 
+		// Document page.
 		_grid_color_button = new ColorButtonVector3();
 		_grid_color_button.value = Vector3(102.0/255.0, 102.0/255.0, 102.0/255.0);
 		_grid_color_button.value_changed.connect(on_color_set);
@@ -73,11 +80,13 @@ public class PreferencesDialog : Gtk.Dialog
 
 		PropertyGrid cv;
 		cv = new PropertyGrid();
+		cv.column_homogeneous = true;
 		cv.add_row(           "Grid", _grid_color_button);
 		cv.add_row("Grid (Disabled)", _grid_disabled_color_button);
 		_document_set.add_property_grid(cv, "Grid");
 
 		cv = new PropertyGrid();
+		cv.column_homogeneous = true;
 		cv.add_row(  "X Axis", _axis_x_color_button);
 		cv.add_row(  "Y Axis", _axis_y_color_button);
 		cv.add_row(  "Z Axis", _axis_z_color_button);
@@ -88,27 +97,47 @@ public class PreferencesDialog : Gtk.Dialog
 		_gizmo_size_spin_button.value_changed.connect(on_gizmo_size_value_changed);
 
 		cv = new PropertyGrid();
+		cv.column_homogeneous = true;
 		cv.add_row("Size", _gizmo_size_spin_button);
 		_document_set.add_property_grid(cv, "Gizmo");
 
+		// Level page.
 		_level_autosave_spin_button = new EntryDouble(5, 1, 60);
 		_level_autosave_spin_button.value_changed.connect(on_level_autosave_value_changed);
 
 		cv = new PropertyGrid();
+		cv.column_homogeneous = true;
 		cv.add_row("Autosave (mins)", _level_autosave_spin_button);
 		_viewport_set.add_property_grid(cv, "Level");
 
+		// Memory and limits page.
 		_log_delete_after_days = new EntryDouble(10, 0, 90);
 		_console_max_lines = new EntryDouble(256, 10, 1024);
+
 		cv = new PropertyGrid();
+		cv.column_homogeneous = true;
 		cv.add_row("Delete logs older than (days)", _log_delete_after_days);
 		cv.add_row("Console max lines", _console_max_lines);
 		_system_set.add_property_grid(cv, "Memory and Limits");
 
+		// External tools page.
+		_lua_external_tool_button = new AppChooserButton("text/plain");
+		_lua_external_tool_button.show_dialog_item = true;
+		_image_external_tool_button = new AppChooserButton("image/*");
+		_image_external_tool_button.show_dialog_item = true;
+
+		cv = new PropertyGrid();
+		cv.column_homogeneous = true;
+		cv.add_row("External Lua editor", _lua_external_tool_button);
+		cv.add_row("External image editor", _image_external_tool_button);
+		_external_tools_set.add_property_grid(cv, "External Editors");
+
+		// Add pages.
 		_notebook = new Gtk.Notebook();
 		_notebook.append_page(_document_set, new Gtk.Label("Document"));
 		_notebook.append_page(_viewport_set, new Gtk.Label("Viewport"));
 		_notebook.append_page(_system_set, new Gtk.Label("System"));
+		_notebook.append_page(_external_tools_set, new Gtk.Label("External Tools"));
 		_notebook.vexpand = true;
 		_notebook.show_border = false;
 
@@ -150,6 +179,46 @@ public class PreferencesDialog : Gtk.Dialog
 		_level_autosave_spin_button.value = preferences.has_key("autosave_timer") ? (double)preferences["autosave_timer"] : _level_autosave_spin_button.value;
 		_log_delete_after_days.value      = preferences.has_key("log_expiration") ? (double)preferences["log_expiration"] : _log_delete_after_days.value;
 		_console_max_lines.value          = preferences.has_key("console_max_lines") ? (double)preferences["console_max_lines"] : _console_max_lines.value;
+
+		// External tools.
+		Hashtable external_tools = preferences.has_key("external_tools")
+			? (Hashtable)preferences["external_tools"]
+			: new Hashtable()
+			;
+
+		Hashtable lua_editor = external_tools.has_key("lua_editor")
+			? (Hashtable)external_tools["lua_editor"]
+			: new Hashtable()
+			;
+
+		string app = "";
+		string app_id = "";
+
+		if (lua_editor.has_key("app"))
+			app = (string)lua_editor["app"];
+		else
+			app = AppChooserButton.APP_DEFAULT;
+
+		if (lua_editor.has_key("app_id"))
+			app_id = (string)lua_editor["app_id"];
+		else
+			app_id = null;
+		_lua_external_tool_button.set_app(app, app_id);
+
+		Hashtable image_editor = external_tools.has_key("image_editor")
+			? (Hashtable)external_tools["image_editor"]
+			: new Hashtable()
+			;
+		if (image_editor.has_key("app"))
+			app = (string)image_editor["app"];
+		else
+			app = AppChooserButton.APP_DEFAULT;
+
+		if (image_editor.has_key("app_id"))
+			app_id = (string)image_editor["app_id"];
+		else
+			app_id = null;
+		_image_external_tool_button.set_app(app, app_id);
 	}
 
 	public void save(Hashtable preferences)
@@ -164,6 +233,24 @@ public class PreferencesDialog : Gtk.Dialog
 		preferences["autosave_timer"] = _level_autosave_spin_button.value;
 		preferences["log_expiration"] = _log_delete_after_days.value;
 		preferences["console_max_lines"] = _console_max_lines.value;
+
+		// External tools.
+		string app;
+		string? app_id;
+		Hashtable lua_editor = new Hashtable();
+		app = _lua_external_tool_button.selected_app(out app_id);
+		lua_editor["app"] = app;
+		lua_editor["app_id"] = app_id != null ? app_id : "";
+
+		Hashtable image_editor = new Hashtable();
+		app = _image_external_tool_button.selected_app(out app_id);
+		image_editor["app"] = app;
+		image_editor["app_id"] = app_id != null ? app_id : "";
+
+		Hashtable external_tools = new Hashtable();
+		external_tools["lua_editor"] = lua_editor;
+		external_tools["image_editor"] = image_editor;
+		preferences["external_tools"] = external_tools;
 	}
 
 	public void apply()